[EC-508] Move DeleteOrganizationUserCommand to OrganizationFeatures folder
