+++ bitwarden_license/src/Scim/Utilities/ScimServiceCollectionExtensions.cs
@@ -1,7 +1,7 @@
-﻿using Bit.Core.Commands;
-using Bit.Core.Commands.Interfaces;
-using Bit.Scim.Queries.Users;
-using Bit.Scim.Queries.Users.Interfaces;
+﻿using Bit.Core.OrganizationFeatures.OrganizationUsers;
+using Bit.Core.OrganizationFeatures.OrganizationUsers.Interfaces;
+using Bit.Scim.Users;
+using Bit.Scim.Users.Interfaces;
 
 namespace Bit.Scim.Utilities;
